Add basic functionality to mach profile_helper script
- Runs colorscheme script
- Runs .sh hooks
- Change hook returns to exit since the scripts aren't sourced

+++ src/main.rs
@@ -1,15 +1,10 @@
 use std::env;
-// use clap::Parser;
-use std::path::{Path,PathBuf};
-use std::fs::{self,File};
-use std::io;
 use std::fmt;
-
-// #[derive(Parser)]
-// struct Cli {
-//     theme_name: String,
-// }
-//
+use std::fs::{self,File};
+use std::io::{self,Read};
+use std::os::unix::fs as unix_fs;
+use std::path::{Path,PathBuf};
+use std::process::Command;
 
 struct PathError {
     message: String,
@@ -22,21 +17,24 @@ impl fmt::Display for PathError {
     }
 }
 
+// Creates a file if it does not exist
 fn ensure_file_exists<P: AsRef<Path>>(file_path: P, is_directory: Option<bool>) -> io::Result<()> {
     let path = file_path.as_ref();
 
     if is_directory == Some(true) {
         if !path.exists() {
-            // Create the file if it does not exist.
             fs::create_dir_all(path)?;
         };
     } else {
-        File::create(path)?;
+        if !path.exists() {
+            File::create(path)?;
+        };
     }
 
     Ok(())
 }
 
+// Create config files if they don't exist
 fn ensure_config_files_exist(base16_config_path: &Path, base16_shell_theme_name_path: &Path) -> Result<(), PathError> {
     ensure_file_exists(base16_config_path, Some(true)).map_err(|_| PathError {
         message: "Failed to create directory. Check if parent directory exists.".to_string(),
@@ -50,18 +48,103 @@ fn ensure_config_files_exist(base16_config_path: &Path, base16_shell_theme_name_
     Ok(())
 }
 
-fn create_env_vars(config_path: &Path) -> (PathBuf, PathBuf, PathBuf, PathBuf) {
-    let base16_config_path = config_path.join("tinted-theming");
-    let base16_shell_path: PathBuf = env::current_dir().expect("Failed to get current directory");
-    let base16_shell_colorscheme_path = base16_config_path.join("base16_shell_theme");
-    let base16_shell_theme_name_path = base16_config_path.join("theme_name");
+// Convert file contents to string
+fn read_file_to_string(path: &Path) -> io::Result<String> {
+    let mut file = File::open(path)?;
+    let mut contents = String::new();
 
-    (base16_config_path, base16_shell_path, base16_shell_colorscheme_path, base16_shell_theme_name_path)
+    file.read_to_string(&mut contents)?;
+
+    Ok(contents)
+}
+
+fn set_theme(
+    theme_name: &String,
+    base16_config_path: &Path,
+    base16_shell_path: &Path,
+    theme_script_path: &Path,
+    base16_shell_colorscheme_path: &Path,
+    base16_shell_theme_name_path: &Path
+) -> std::io::Result<()> {
+    let current_theme_name = match read_file_to_string(base16_shell_theme_name_path) {
+        Ok(contents) => contents,
+        Err(e) => {
+            eprintln!("Failed to read from file: {}", e);
+            return Err(e);
+        }
+    };
+
+    if theme_name.to_string() == current_theme_name {
+        eprintln!("Theme \"{}\" is already set", theme_name);
+        std::process::exit(1);
+    }
+
+    // Remove symlink file and create colorscheme symlink
+    // -----------------------------------------------------------------
+    if base16_shell_colorscheme_path.exists() {
+        fs::remove_file(base16_shell_colorscheme_path)?;
+    }
+
+    unix_fs::symlink(theme_script_path, base16_shell_colorscheme_path)?;
+
+    // Write theme name to file
+    // -----------------------------------------------------------------
+    fs::write(base16_shell_theme_name_path, theme_name)?;
+
+    // Run colorscheme script
+    // -----------------------------------------------------------------
+
+    // Source colorscheme script
+    // Wait for script to fully execute before continuing
+    let mut child = Command::new("/bin/bash").arg(base16_shell_colorscheme_path).spawn()?;
+    let status = child.wait()?;
+    if !status.success() {
+        eprintln!("Command finished with a non-zero status.");
+        return Err(
+            std::io::Error::new(
+                std::io::ErrorKind::Other,
+                "Script execution failed"
+            )
+        );
+    }
+
+    // Hooks
+    // Set env variables for hooks and then execute .sh hooks
+    // -----------------------------------------------------------------
+    let mut base16_shell_hooks_path = env::var("BASE16_SHELL_HOOKS_PATH").unwrap_or_default();
+    if base16_shell_hooks_path.is_empty() || !Path::new(&base16_shell_hooks_path).is_dir() {
+        base16_shell_hooks_path = format!("{}/hooks", base16_shell_path.display());
+
+        env::set_var("BASE16_SHELL_HOOKS_PATH", &base16_shell_hooks_path);
+    }
+
+    env::set_var("BASE16_SHELL_THEME_NAME_PATH", base16_shell_theme_name_path);
+    env::set_var("BASE16_CONFIG_PATH", base16_config_path);
+
+    let base16_shell_hooks_path = base16_shell_path.join("hooks");
+    if base16_shell_hooks_path.is_dir() {
+        for entry in fs::read_dir(base16_shell_hooks_path)? {
+            let entry = entry?;
+            let path = entry.path();
+
+            // Check if the file name ends with .sh
+            if path.extension().and_then(|ext| ext.to_str()) == Some("sh") {
+                Command::new("/bin/bash").arg(path).status()?;
+            }
+        }
+    }
+
+    Ok(())
 }
 
 fn main() {
-    // let args = Cli::parse();
-    let exe_path: Option<PathBuf> = env::current_exe().ok().and_then(|path| path.canonicalize().ok());
+    let args: Vec<String> = env::args().collect();
+
+    if args.len() < 2 {
+        eprintln!("You didn't provide a theme name argument.");
+        std::process::exit(1);
+    }
+
     let config_path: PathBuf = env::var("XDG_CONFIG_HOME")
         .map(PathBuf::from)
         .unwrap_or_else(|_| {
@@ -69,65 +152,37 @@ fn main() {
 
             PathBuf::from(home).join(".config")
         });
+    let config_path_something = config_path.as_path();
+    let base16_config_path = config_path_something.join("tinted-theming");
+    let base16_shell_path: PathBuf = env::current_dir().expect("Failed to get current directory");
+    let base16_shell_colorscheme_path = base16_config_path.join("base16_shell_theme");
+    let base16_shell_theme_name_path = base16_config_path.join("theme_name");
+    let theme_name = &args[1];
+    let theme_script_path = base16_shell_path.join(format!("scripts/base16-{}.sh", theme_name));
 
-    let (base16_config_path, base16_shell_path, base16_shell_colorscheme_path, base16_shell_theme_name_path) = create_env_vars(config_path.as_path());
-
-    // let base16_config_path = config_path.join("tinted-theming-test");
-    // let base16_shell_colorscheme_path = base16_config_path.join("base16_shell_theme");
-    // let base16_shell_theme_name_path = base16_config_path.join("theme_name");
-
-    let base16_shell_path: PathBuf = match env::current_dir() {
-        Ok(path) => path,
-        Err(e) => {
-            println!("Error getting the current directory: {}", e);
-            return;
-        },
-    };
-
-
-    // if let Some(script_path) = exe_path {
-    //     let base16_shell_colorscheme_path_str = base16_shell_colorscheme_path.to_str().expect("invalid path");
-
-    //     println!(
-    //         "Config path: {} {} {}", 
-    //         base16_shell_colorscheme_path_str,
-    //         script_path.display(),
-    //         base16_shell_theme_name_path.display()
-    //     );
-    // }
-
-
-    let mut base16_shell_hooks_path = env::var("BASE16_SHELL_HOOKS_PATH").unwrap_or_default();
-
-    if base16_shell_hooks_path.is_empty() || !Path::new(&base16_shell_hooks_path).is_dir() {
-        base16_shell_hooks_path = format!("{}/hooks", base16_shell_path.display());
-
-        env::set_var("BASE16_SHELL_HOOKS_PATH", &base16_shell_hooks_path);
+    if !theme_script_path.exists() {
+        eprintln!("Theme \"{}\" does not exist, try a different theme", theme_name);
+        std::process::exit(1);
     }
-
-    // Use base16_shell_hooks_path as needed
 
     if let Err(e) = ensure_config_files_exist(base16_config_path.as_path(), base16_shell_theme_name_path.as_path()) {
         eprintln!("Error: {}", e);
     }
 
-
-    println!("script path: {}", base16_shell_path.display());
-    println!("BASE16_SHELL_HOOKS_PATH is {}", base16_shell_hooks_path);
-    println!("base16_shell_path is {}", base16_shell_path.display());
-    println!("base16_config_path is {}", base16_config_path.display());
-    println!("base16_shell_colorscheme_path is {}", base16_shell_colorscheme_path.display());
-    println!("base16_shell_path is {}", base16_shell_path.display());
-
-
-    // let user_config_dir_path = user_home_dir_path.unwrap_or_else(|_| default_value.to_string());
-    // let base_path = env::var("XDG_CONFIG_HOME")
-    //     .or_else(|_| env::var("HOME"))
-    //     .unwrap_or_else(|_| panic!("Neither XDG_CONFIG_HOME nor HOME environment variable is set"));
-
-    // // Append '/cli_app' to the base path.
-    // let config_path = format!("{}/cli_app", base_path);
-
-
-    // print!("{username}");
+    match set_theme(
+        &theme_name, 
+        base16_config_path.as_path(),
+        base16_shell_path.as_path(),
+        theme_script_path.as_path(),
+        base16_shell_colorscheme_path.as_path(),
+        base16_shell_theme_name_path.as_path()
+    ) {
+        Ok(()) => {
+            eprintln!("Theme set to: {}", theme_name);
+        },
+        Err(e) => {
+            eprintln!("Failed to set theme \"{}\": {}", theme_name, e);
+            std::process::exit(1);
+        }
+    }
 }
